Move playSound function to its own module
